+++ packages/web/package.json
@@ -13,7 +13,7 @@
         "react": "^18.2.0",
         "react-bootstrap": "^2.4.0",
         "react-dom": "^18.2.0",
-        "react-hook-form": "^7.43.4",
+        "react-hook-form": "^7.49.2",
         "react-hot-toast": "^2.4.1",
         "react-hovering-cards-carousel": "^0.9.71",
         "react-icons": "^4.6.0",
